requirements.txt
pika==1.2.0
mysql-connector-python==8.0.26
requests==2.26.0
